requirements.txt
Flask>=2.2
camelot-py[cv]
pandas
openpyxl
